+++ CHANGES.rst
@@ -17,7 +17,7 @@ Cubeviz
 Imviz
 ^^^^^
 
-- Footprints plugin now supports selecting the spatially closest overlay
+- Footprints plugin now supports selecting the closest overlay
   to a clicked point in the image viewer. [#3525]
 
 Mosviz